Move all authors into same namespace
