+++ src/Firehose.Web/Authors/DavidBritch.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Firehose.Web
+namespace Firehose.Web.Authors
 {
     public class DavidBritch : IWorkAtXamarinOrMicrosoft
 	{
