+++ src/Firehose.Web/Authors/DevlinDuldulao.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using Firehose.Web.Infrastructure;
 
-namespace Firehose.Web
+namespace Firehose.Web.Authors
 {
     public class DevlinDuldulao : IAmACommunityMember
     {
