+++ src/Firehose.Web/Authors/JonDick.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using Firehose.Web.Infrastructure;
 
-namespace Firehose.Web
+namespace Firehose.Web.Authors
 {
     public class JonDick : IWorkAtXamarinOrMicrosoft
     {
