+++ src/Firehose.Web/Authors/MatthewLeibowitz.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using Firehose.Web.Infrastructure;
 
-namespace Firehose.Web
+namespace Firehose.Web.Authors
 {
     public class MatthewLeibowitz : IWorkAtXamarinOrMicrosoft
     {
